.claude/worktrees/intelligent-napier/pyproject.toml
[project]
name = "voice-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "elevenlabs>=2.41.0",
    "generativeai>=0.0.1",
    "google-genai>=1.70.0",
    "openai>=2.30.0",
    "pyaudio>=0.2.14",
    "pydub>=0.25.1",
    "python-dotenv>=1.2.2",
    "speechrecognition>=3.16.0",
    "streamlit>=1.56.0",
    "streamlit-extras>=1.3.0",
]
